feat(LoadingMessages): add random messages
