+++ frontend/src/components/Video/VideoPlayer/LoadingMessage/LoadingMessage.module.css
@@ -1,0 +1,71 @@
+.loadingMessage {
+  /* font-family: "Roboto", sans-serif; */
+  font-family: inherit;
+  font-size: 2.5rem;
+  display: inline-block;
+}
+
+.character {
+  animation: loading infinite linear;
+}
+
+@keyframes loading {
+  0% {
+    translate: 0 0;
+  }
+  25% {
+    translate: 0 0.1rem;
+  }
+  75% {
+    translate: 0 -0.1rem;
+  }
+  100% {
+    translate: 0 0;
+  }
+}
+
+@media (max-width: 1250px) {
+  .loadingMessage {
+    font-size: 2rem;
+  }
+  @keyframes loading {
+    0% {
+      translate: 0 0;
+    }
+    25% {
+      translate: 0 0.05rem;
+    }
+    75% {
+      translate: 0 -0.05rem;
+    }
+    100% {
+      translate: 0 0;
+    }
+  }
+}
+
+@media (max-width: 400px) {
+  .loadingMessage {
+    font-size: 1.7rem;
+  }
+  @keyframes loading {
+    0% {
+      translate: 0 0;
+    }
+    25% {
+      translate: 0 0.03rem;
+    }
+    75% {
+      translate: 0 -0.03rem;
+    }
+    100% {
+      translate: 0 0;
+    }
+  }
+}
+
+@media (max-width: 320px) {
+  .loadingMessage {
+    font-size: 1.4rem;
+  }
+}
